Fix bugs during upgrade 0.94 to 0.98
Summary: Ref T5833

Test Plan: N/A

Reviewers: fenghonghua, heliangliang, chengqiming, wuzesheng, zhouyingchao, zhangguanghao1, cuijianwei

Reviewed By: cuijianwei

Subscribers: infra-dev

Maniphest Tasks: T5833

Differential Revision: https://phabricator.d.xiaomi.net/D29247

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/migration/NamespaceUpgrade.java
@@ -334,8 +334,12 @@ public class NamespaceUpgrade implements Tool {
     removeTableInfoInPre96Format(TableName.META_TABLE_NAME);
 
     Path oldRootDir = new Path(rootDir, "-ROOT-");
-    if(!fs.rename(oldRootDir, backupDir)) {
-      throw new IllegalStateException("Failed to old data: "+oldRootDir+" to "+backupDir);
+    if (!fs.exists(oldRootDir) && fs.exists(new Path(backupDir, "-ROOT-"))) {
+      LOG.info("Root dir has been moved to " + backupDir);
+    } else {
+      if (!fs.rename(oldRootDir, backupDir)) {
+        throw new IllegalStateException("Failed to old data: " + oldRootDir + " to " + backupDir);
+      }
     }
   }
